Moved ClientLoop class into Client
