+++ jenjin-core-client/src/main/java/com/jenjinstudios/client/net/Client.java
@@ -5,10 +5,14 @@ import com.jenjinstudios.core.MessageIO;
 import com.jenjinstudios.core.io.Message;
 import com.jenjinstudios.core.io.MessageRegistry;
 
+import java.io.IOException;
 import java.security.KeyPair;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Timer;
+import java.util.TimerTask;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 /**
  * The base class for any client.  This class uses a similar system to the JGSA.
@@ -97,4 +101,65 @@ public class Client extends Connection
      * @return The average number of updates per second that this client is executing.
      */
     public double getAverageUPS() { return 1.0d / clientLoop.getAverageRunTime(); }
+
+    /**
+     * The ClientLoop class is essentially what amounts to the output thread.
+     *
+     * @author Caleb Brinkman
+     */
+
+    private static class ClientLoop extends TimerTask
+    {
+        private static final int MAX_STORED_UPDATE_TIMES = 50;
+        private static final int NANOS_TO_SECOND = 1000000000;
+        private static final Logger LOGGER = Logger.getLogger(ClientLoop.class.getName());
+        /** The Client for this loop. */
+        private final Client client;
+        private int updateCount;
+        private long lastStart = System.nanoTime();
+        private final long[] updateTimesNanos = new long[MAX_STORED_UPDATE_TIMES];
+
+        /**
+         * Construct a ClientLoop for the given client.
+         *
+         * @param client The client for this ClientLoop
+         */
+        ClientLoop(Client client) {
+            this.client = client;
+        }
+
+        @Override
+        public void run() {
+            updateCount++;
+            saveUpdateTime();
+            client.runRepeatedTasks();
+            client.getExecutableMessageQueue().runQueuedExecutableMessages();
+            try
+            {
+                client.getMessageIO().writeAllMessages();
+            } catch (IOException e)
+            {
+                LOGGER.log(Level.INFO, "Shutting down because of IOException", e);
+                client.shutdown();
+            }
+        }
+
+        private void saveUpdateTime() {
+            long newStart = System.nanoTime();
+            long timeElapsed = newStart - lastStart;
+            lastStart = newStart;
+            updateTimesNanos[updateCount % updateTimesNanos.length] = timeElapsed;
+        }
+
+        public double getAverageRunTime() {
+            double maxIndex = Math.min(updateCount, updateTimesNanos.length);
+            double total = 0;
+            for (int i = 0; i < maxIndex; i++)
+            {
+                total += updateTimesNanos[i];
+            }
+            return total / maxIndex / NANOS_TO_SECOND;
+        }
+
+    }
 }
